[fix bug] 10**decimal give invalid result

+++ contracts/TokenA.sol
@@ -32,7 +32,7 @@ contract TokenA {
     }
 
     function transfer(address receiver, uint numTokens) public returns (bool) {
-        uint256 value = numTokens.mul(10 ** this.decimals());
+        uint256 value = numTokens.mul(10 ** uint256(this.decimals()));
         require(receiver != address(0));
         require(value <= balances[msg.sender]);
         balances[msg.sender] = balances[msg.sender].sub(value);
@@ -44,7 +44,7 @@ contract TokenA {
     }
 
     function approve(address _spender, uint numTokens) public returns (bool) {
-        uint256 value = numTokens.mul(10 ** this.decimals());
+        uint256 value = numTokens.mul(10 ** uint256(this.decimals()));
         allowed[_spender][msg.sender] = value;
         emit Approval(_spender ,msg.sender, value);
         return true;
@@ -55,7 +55,7 @@ contract TokenA {
     }
 
     function transferFrom(address owner, address buyer, uint numTokens) public returns (bool) {
-        uint256 value = numTokens.mul(10 ** this.decimals());
+        uint256 value = numTokens.mul(10 ** uint256(this.decimals()));
         require(value <= balances[owner]);
         require(value <= allowed[owner][msg.sender]);
